Test anonymous get on privileged resource returns redirect
